# Conflicts:
#	src/main/java/com/iu/start/bankbook/BookController.java
#	src/main/java/com/iu/start/bankmembers/BankMembersDAO.java
#	src/main/java/com/iu/start/bankmembers/BankMembersFileDTO.java
#	src/main/java/com/iu/start/bankmembers/MemberController.java
#	src/main/java/com/iu/start/bankmembers/MembersService.java
#	src/main/java/com/iu/start/board/notice/NoticeController.java
#	src/main/java/com/iu/start/board/notice/NoticeService.java
#	src/main/java/com/iu/start/file/FileDTO.java
#	src/main/resources/database/config/mybatisConfig.xml
#	src/main/resources/database/mappers/BankMembersMapper.xml
#	src/main/webapp/WEB-INF/views/member/join.jsp
#	src/main/webapp/WEB-INF/views/member/login.jsp
#	src/main/webapp/WEB-INF/views/member/myPage.jsp

+++ src/main/java/com/iu/start/bankaccount/BankAccountDTO.java
@@ -6,7 +6,7 @@ import com.iu.start.bankbook.BankBookDTO;
 
 public class BankAccountDTO {
 	
-	private Integer accountNum;
+	private Long accountNum;
 	private String id;
 	private Long bookNum;
 	private Date accountDate;
@@ -20,11 +20,11 @@ public class BankAccountDTO {
 		this.bankBookDTO = bankBookDTO;
 	}
 
-	public Integer getAccountNum() {
+	public Long getAccountNum() {
 		return accountNum;
 	}
 	
-	public void setAccountNum(Integer accountNum) {
+	public void setAccountNum(Long accountNum) {
 		this.accountNum = accountNum;
 	}
